# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateOrderRequest:
    properties:
      payment_method:
        type: string
      schedule_id:
        type: integer
      seats:
        items:
          type: integer
        type: array
    required:
    - payment_method
    - schedule_id
    - seats
    type: object
...
  dto.GetSchedules:
    properties:
      cinema_city:
        type: string
      cinema_id:
        type: integer
      cinema_location:
        type: string
      cinema_logo:
        type: string
      cinema_name:
        type: string
      id:
        type: integer
      price:
        type: integer
      show_date:
        type: string
      show_time:
        type: string
    type: object
...
  dto.PaginationMeta:
    properties:
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      total_page:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      msg:
        type: string
      success:
        type: boolean
    type: object
  dto.SeatResponse:
    properties:
      row_letter:
        type: string
      seat_id:
        type: integer
      seat_number:
        type: integer
      seat_type:
        type: string
      status:
        type: string
    type: object
  dto.UpdateOrderRequest:
    properties:
      payment_status:
        type: string
    required:
    - payment_status
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Back End using go with gin.
  title: Tickitz Back End
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      description: Get movies with search, genre filter, and pagination
      parameters:
      - description: Search by title
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by genre ID
        in: query
        items:
          type: integer
        name: genre_id
        type: array
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Filter movies
      tags:
      - movies
  /movies/detail/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific movie
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get movie detail
      tags:
      - movies
  /movies/popular:
    get:
      consumes:
      - application/json
      description: Get a list of popular movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get popular movies
      tags:
      - movies
  /movies/upcoming:
    get:
      consumes:
      - application/json
      description: Get a list of upcoming movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get upcoming movies
      tags:
      - movies
  /orders:
    post:
      consumes:
      - application/json
      description: Create a new ticket order for a user (Requires user token)
      parameters:
      - description: Order Body
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    patch:
      consumes:
      - application/json
      description: Update the payment status of an order (Requires admin token)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status Body
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update order payment status
      tags:
      - orders
  /orders/schedules/{id}:
    get:
      consumes:
      - application/json
      description: Get list of schedules with cinema details filtered by movie ID,
        date, and city
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Show Date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: City Name
        in: query
        name: city
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetSchedules'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get schedules for a movie
      tags:
      - orders
  /orders/seats/{id}:
    get:
      consumes:
      - application/json
      description: Get list of seats status (sold/available) for a specific schedule
        ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SeatResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get seats for a schedule
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"